Add non static functions: compile() and eval()
With this commit the class can now be instantiated.
It now has a constructor and a destructor.

The Compile function will read an expression,
produce an RPN, and save it on this->RPN.

Eval will evaluate this->RPN only when asked for.

The calculator::str() function was added to simplify
the debugging process.