Modify calculate() for lazy variable evaluation
Now it is possible to precompile an expression without
binding all its variable values. And bind them only
when the expression will be evaluated.

This way different evaluations of a single expression
may have different sets of variables.

This method optimizes the case when you need to re-evaluate
a single equation several times in different contexts.