Add the Scope concept and the Scope Class
Now instead of passing only maps of variables (TokenMap_t*)
to calculator instances you can keep record a Scope hierarchy.

The scope class works as a stack of TokenMap_t*.

When searching for a variable, it searches the
stack on a top-down order and stops as soon as it finds
a match for the search key.

The old API was not modified, everything still works as before,
because the Scope class constructor accepts a TokenMap_t*, so
you can insert a TokenMap_t* anywhere you could use an Scope instance.