Make packToken(TokenBase*) less error prone
packToken(TokenBase*) not only builds a packToken but also
gets responsible for dellocating it, and this is not a normal
class behavior, since it could lead to a error prone code like this:

int main() {
  TokenBase* b = new Token<double>(10, NUM);
  packToken p = b;

    ... some code here ...

  delete b;

    ... some more code ...

  return 0; // Seg fault.
}

This commit makes this function private and explicit,
and only accessible from inside the class calculator which is
now friend of class packToken.

It was done to guarantee that only main developers have need
to fully comprehend this constructor behavior avoiding headaches.