Add support for assignment on existing variables
Now when resolving a variable's value it is created an rvalue
Token which keeps a reference to the original variable name.

This Token is then qualified to be used on the left side of
assigment operations.

When this is not the case it is converted to a normal Token
so that it can be used for calculations.