Make `toRPN` ignore delimiters if between brackets
Before this commit using the calculator class as a slave parser
would not work as expected if the delimiter were a ')' or a '\n'.

In other words the examples below would not work:

  const char delim = ")";
  const char* text = " if ( (a+b)*c == 3 ) { ... }";
  c.compile(text+5, delim);

  const char delim = "\n";
  const char* text = " a = (\n  1,\n  2,\n  3\n)\n more_code(...)\n ";
  c.compile(text+5, delim);

These examples would fail because the parser would stop on the
first ocurrence of their respective delimiters.

Now the delimiters are only considered when not between brackets.