Make packToken(TokenBase*) public
It was set as private because it should be used with care,
since an unaware user might forget that the TokenBase* memory
will be deleted once the packToken is destroied.

But since this is a necessary feature of the API it is better
to stay public and with the `explicit` modifier on it.