Change assignment rules to protect global scope
TLDR; The global scope is special and should be protected.

The user should not be able to implicitly overwrite a variable if he
did not declare it.

It's not feasible to expect our users to remember all variables
that exist on the global scope and avoid using their names on
assignments. So now if a user try to assign a value to a built-in
function on global scope, it will instead assign it to the local scope.

The user is still capable of writting stuff on non-local variables,
as long as these variables scope is not the default global scope.