Simplify containers (#26)
* Add a new CppFunction constructor for funcs with 0 args

* Update `pack` class API

* Remove `packMap` type in favor of the `TokenMap` class

The `packMap` type was defined as a `pack<TokenMap>`, so it would
manage memory of all Maps used on the system.

TLDR; This is a better design choice than it was before.

The reason for this change is to solve this 2 problems:

1. It creates 2 maps that do almost the same thing: packMap and TokenMap
   and it might confuse a reader of the code.

2. It could not count all TokenMap references, since some of these
   maps were referenced from inside the TokenMap class, i.e. the
   parent TokenMap of each TokenMap.

For those reasons we modified the TokenMap class. Now it inherits
from the pack<> class and manage a reference to a struct called `MapData_t`

----

This commit also makes TokenMap inherit from TokenBase, as well as
the Iterable interface.

This way it will be possible to cast a TokenBase* to an Iterable*
if the token type is an iterator, i.e. if `IT & type != 0`.

* Remove `packList` type in favor of `TokenList` class

Now TokenList inherits from:
- pack<TokenList_t>
- TokenBase
- Iterable

The reasons for this commit are the same of the last commit.

* Change "arglist" function keyword to "args"

* Refactor: Iterable is now a public TokenBase

Now there is no need to make TokenMap and TokenList
inherit from TokenBase and Iterable.

Now the Iterable class means "an Iterable Token"

* Make Iterator inherit from TokenBase

* Fix Travis build