Fix bugs with sub-parsers on `toRPN()` function
Before this commit if a sub-parser left an empty
expression it would cause a segmentation fault.

Also the `LineComment` sub-parser defined on the `builtin-features.cpp`
file would consume the '\n' on the end of the expression, making
it impossible for this character to be used as delimiter by the
calculator. The updated version fixes it.