Copy `map` and `list` constructors to `objects.*`
These constructors were defined as CppFunctions on
`builtin-features.cpp`, now that we are defining
the implicit constructors for map: `{}` and list: `[]` it
was necessary to define it on someplace the user could not change.

The original code on `builtin-features.cpp` is unchanged. The user
can still overwrite his own definition of the builtin `map` and `list`
functions, as he wish.