A lot of refactors, some optimizations and `{}` and `[]` JSON constructors (#44)
* Simplify `handle_op()` and `handle_unary()` functions

Now both receive an rpnBuilder as argument, instead of receiving
several different variables.

`handle_unary()` now writes to `data->lastTokenWasUnary()` instead
of returning a true/false argument to be written on it outside
the function as it was before. The new return type of the
function is void.

* Join the functions `handle_unary` and `handle_op`

Since they were always called together and did almost the same
thing there was no reason to keep them separated.

* Minor refactor - Move `opPrecedence` into `rpnData`

Now the function `handle_op` requires only 2 arguments:

- const char* expr
- rpnBuilder* data

* Expose `handle_op` to API

`handle_op()` function was exposed to API so that it could be used
on custom reserved word parsers defined by our users.

* Add the builtin map constructor operator: '{}'

Now it is possible to create maps just by writing:

```C++
  calculator::calculate("M = { 'my':'map' }");
```

* Copy `map` and `list` constructors to `objects.*`

These constructors were defined as CppFunctions on
`builtin-features.cpp`, now that we are defining
the implicit constructors for map: `{}` and list: `[]` it
was necessary to define it on someplace the user could not change.

The original code on `builtin-features.cpp` is unchanged. The user
can still overwrite his own definition of the builtin `map` and `list`
functions, as he wish.

* Add list implicit constructor `[1,2,3]`

Now you can declare a new list as you would in JavaScript:

```C++
  calculator::calculate("L = [1,2,3]");
```

* Add logo to repository

* Add move constructor to packToken class

* Move some functions into the rpnBuilder class

The functions:

- `handle_op()`
- `handle_unary()`
- `cleanRPN()`

were moved inside the rpnBuilder class, allowing now to
use `data.handle_op("+")` instead of `handle_op("+", &data)`.

More functions and functionalities should be moved into this class
in future, since we can use it to decouple the rpn building process
from the calculator class.

* Minor refactor: Move open/close bracket code to functions

* Minor refactor - move 2 functions in `builtin-features.cpp`

* Make public the static functions: toRPN and calculate

These 2 functions are useful to be used on the user defined
reserved word parsers, as such they are now available
to be called outside the calculator class.

* Refactor: Decouple internal variables of calculate()

Now the internal variables of the calculate function are
joined together on a struct called "evaluationData".

It can then be exposed for the user to access it in future commits.

* Refactor: Expose calculator variables to Operation::exec()

Now the `Operation::exec` function's signature is:

```C++
  packToken exec(const packToken& left, const packToken& right,
                 const evaluationData* data);
```

The std::string op (that was before this commit the second argument
of the exec() function) is now available inside the evaluationData type.

This struct also contains several other internal variables used
by the calculator::calculate() function. This commit thus allows
more power and flexibility to user Operations.