Make possible for functions to work as left operands
Before this commit the calculator would try to execute the
hard-coded operator `()` whenever the left operand was a
function, thus causing an error if the operator was not
in fact the call operator `()`:

```C++
calculator::calculate("pow, 'any thing else'");
// Undefined operation with operands: [Function: pow] and 'any thing else'
```

Now the operator `()` will only be called when it is the actual
operator being called. Thus allowing other operators to work
with functions as their left operand.