Op Associativity API + Minor fixes, refactors and updates (#45)
* Add error check for TokenList::operator[]

* Add `push` and `pop` functions to TokenList class

Now it is possible to use the TokenList class without accessing
its internal `std::vector<packToken>` as it was before.

* Make possible for functions to work as left operands

Before this commit the calculator would try to execute the
hard-coded operator `()` whenever the left operand was a
function, thus causing an error if the operator was not
in fact the call operator `()`:

```C++
calculator::calculate("pow, 'any thing else'");
// Undefined operation with operands: [Function: pow] and 'any thing else'
```

Now the operator `()` will only be called when it is the actual
operator being called. Thus allowing other operators to work
with functions as their left operand.

* Expose API for setting OP associativity

Before this commit you would register an operator like this:

```C++
OppMap_t& opp = calculator::default_opPrecedence();
opp["+"] = 2;
```

And it was not possible to define the operator associativity.
Now it is possible, for that you just need to define the
operator precedence with a negative sign for Right to Left
precedence, meaning Left to Right otherwise, e.g.:

```C++
OppMap_t& opp = calculator::default_opPrecedence();
opp.add("+", 2);   // Left to right associativity
opp.add("=", -2);  // Right to left associativity
```

* Make `handle_op` to set `lastTokenWasOp` flag

* Add `rpnBuilder::handle_token(TokenBase*)` function

This function is responsible for adding a token to the rpn queue
and also setting the flags `lastTokenWasOp` and `lastTokenWasUnary`
to false. So that our user will not have to handle these flags anymore.

* Simplify `toRPN` detection of the operator `()`