Fix static initialization order bug with packToken::None
Before this commit there was a hidden bug on packToken::None.

If it was called from a static initialized constructor it might
or might not cause a bug.

Now packToken::None is no longe a packToken, but instead a function
that returns a packToken when called.

To adapt old code to conform with this change you should replace
any occurency of `packToken::None` for `packToken::None()`.