Group together: OppMap_t, rWordMap_t and opMap_t
Now these 3 containers were grouped into a a bigger one: Config_t

This change was done for all the 3 contained configuration data.
And this data had to be passed along on many functions.

Now only 1 argument is required to be passed.

An extra reason is for when a user wants to create his own calculator,
Before this he would need to overwrite 3 virtual functions, one
for each of those configuration containers, now 1 is enough.

The old static variables:

- calculator::default_opPrecedence
- calculator::default_opMap
- calculator::default_rWordMap

Were replaced by the Default static variable:

- Config_t calculator::Default

To access them now it is possible like this:

- calculator::Default().opPrecedence
- calculator::Default().opMap
- calculator::Default().rWordMap